refactor: changed class name for ellipsoid parameters to datum
