--- doc/source/api_reference/constants.rst
@@ -1,24 +1,0 @@
-=========
-constants
-=========
-
-Gravitational and ellipsoidal parameters [1]_ [2]_
-
-`Source code`__
-
-.. __: https://github.com/tsutterley/pyTMD/blob/main/pyTMD/constants.py
-
-General Attributes and Methods
-==============================
-
-.. autoclass:: pyTMD.constants
-   :members:
-
-References
-----------
-.. [1] B. Hofmann-Wellenhof and H. Moritz, *Physical Geodesy*, 2nd Edition,
-   403 pp., (2006). `doi: 10.1007/978-3-211-33545-1
-   <https://doi.org/10.1007/978-3-211-33545-1>`_
-.. [2] G. Petit and B. Luzum (eds.), *IERS Conventions (2010)*,
-   International Earth Rotation and Reference Systems Service (IERS),
-   `IERS Technical Note No. 36 <https://iers-conventions.obspm.fr/content/tn36.pdf>`_
